Move `snapshots` command to `snapshot list`
